WaveManager.csからEnemy制御のコードを分離しリファクタリングを行った
