+++ Assets/OtherAssets/AnimationValidator/Package/AnimationValidator_v1.0.0.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f6a297bfdceda894c9399e9d131bea29
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
